Refactor: Decouple Memory Consolidation Service into modular pipeline
